+++ react/__tests__/Carousel.test.js
@@ -7,8 +7,41 @@ import Carousel from '../Carousel'
 describe('Carousel component', () => {
   let wrapper
 
+  let props = {
+    height: 440,
+    mobileHeight: 339,
+    showArrows: true,
+    showDots: true,
+    autoplay: true,
+    autoplaySpeed: 4,
+    banner1: {
+      image:
+        'https://raw.githubusercontent.com/vtex-apps/carousel/master/images/banners-01.png',
+      mobileImage:
+        'https://raw.githubusercontent.com/vtex-apps/carousel/master/images/banners-mobile-01.png',
+      page: '/',
+      description: 'banner',
+    },
+    banner2: {
+      image:
+        'https://raw.githubusercontent.com/vtex-apps/carousel/master/images/banners-02.png',
+      mobileImage:
+        'https://raw.githubusercontent.com/vtex-apps/carousel/master/images/banners-mobile-02.png',
+      page: '/',
+      description: 'banner',
+    },
+    banner3: {
+      image:
+        'https://raw.githubusercontent.com/vtex-apps/carousel/master/images/banners-03.png',
+      mobileImage:
+        'https://raw.githubusercontent.com/vtex-apps/carousel/master/images/banners-mobile-03.png',
+      page: '/',
+      description: 'banner',
+    },
+  }
+
   beforeEach(() => {
-    wrapper = render(<Carousel />)
+    wrapper = render(<Carousel {...props} />)
   })
 
   it('should be rendered', () => {